Enable QtMultimedia FFmpeg plugin build in packaging configs
(cherry picked from commit b794125ee9236b48304f29317c42df30ce642b11)

+++ coin/provisioning/common/linux/install-ffmpeg.sh
@@ -49,7 +49,12 @@ source "${BASH_SOURCE%/*}/../unix/SetEnvVar.sh"
 
 ffmpeg_config_options=$(cat "${BASH_SOURCE%/*}/../shared/ffmpeg_config_options.txt")
 
-sudo apt install yasm
+if [ -f /etc/redhat-release ]
+then
+  sudo sudo yum -y install yasm
+else
+  sudo apt install yasm
+fi
 
 version="n5.0"
 ffmpeg_name="FFmpeg-$version"
